+++ src/main/java/nz/co/noirland/noirstore/config/ItemConfig.java
@@ -53,7 +53,7 @@ public class ItemConfig extends Config {
         for(int i = 0;i < (keys.size()-1);i++) {
             int lower = Integer.parseInt(keys.get(i));
             int upper = Integer.parseInt(keys.get(i+1));
-            ret.add(new PriceRange(lower, upper, pricesConfig.getLong(keys.get(i)), pricesConfig.getLong(keys.get(i+1))));
+            ret.add(new PriceRange(lower, upper, pricesConfig.getDouble(keys.get(i)), pricesConfig.getDouble(keys.get(i+1))));
         }
 
         return ret;
